Change PV and PVC status capacity on resize
We add the interface for volume expansion. This will not have any effect
on the underlying volume, but will change properly the PV capacity and
PVC status when a claim is made, and the proper events will be produced
as well.

To do that the external-resizer is used provided by Kubernetes for this
purpose.

Signed-off-by: Marjus Cako <marios.cako@gmail.com>