Repair code example in storageClass description.
I copied this example and had a really rough go as a kubernetes beginner; I needed a custom storageClass for the RabbitMQ operator.

So my pods were broken for a long while, until I changed my storageClass to match the other `local-path` storageClass already enabled in my cluster. This `rancher.io/local-path` seems to be the key to an easy deployment of custom storageClasses.