e2e: confirm the cleanup of PVs with legacy affinity attributes
This applies a small refactor to the e2e tests to ensure that the newer
provisioner is capable of siting helper pods correctly to clean up PVs
with "legacy" affinity constraints.

The kind cluster itself is reconfigured to ensure that all nodes have
`metadata.name` != `metadata.labels["kubernetes.io/hostname"]`, which is
an assumption that does not hold for many cloud providers.